Add stores listing endpoint with filtering
Create api endpoint that returns stores with their according disctances
from a defined location.

Endpoint accepts paging and filtering.
the following filters can be used :
  * type : The type of the desired stores.
  * radius : the maximus distance accepted.

+++ src/Entity/Store.php
@@ -22,12 +22,12 @@ class Store
     private $name;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="decimal")
      */
     private $longitude;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(type="decimal")
      */
     private $latitude;
 
@@ -58,24 +58,24 @@ class Store
         return $this;
     }
 
-    public function getLongitude(): ?int
+    public function getLongitude(): ?float
     {
         return $this->longitude;
     }
 
-    public function setLongitude(int $longitude): self
+    public function setLongitude(float $longitude): self
     {
         $this->longitude = $longitude;
 
         return $this;
     }
 
-    public function getLatitude(): ?int
+    public function getLatitude(): ?float
     {
         return $this->latitude;
     }
 
-    public function setLatitude(int $latitude): self
+    public function setLatitude(float $latitude): self
     {
         $this->latitude = $latitude;
